feat: move command store to command handler, create targets for command_entry and message_entry
