+++ src/agent/centralized_configuration/CMakeLists.txt
@@ -10,7 +10,7 @@ find_package(nlohmann_json REQUIRED)
 
 add_library(CentralizedConfiguration src/centralized_configuration.cpp)
 target_include_directories(CentralizedConfiguration PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
-target_link_libraries(CentralizedConfiguration PUBLIC ModuleCommand Boost::asio nlohmann_json::nlohmann_json Config FilesystemWrapper PRIVATE Logger)
+target_link_libraries(CentralizedConfiguration PUBLIC Boost::asio nlohmann_json::nlohmann_json Config FilesystemWrapper CommandEntry PRIVATE Logger)
 
 include(../../cmake/ConfigureTarget.cmake)
 configure_target(CentralizedConfiguration)
