+++ src/agent/command_handler/tests/CMakeLists.txt
@@ -4,3 +4,8 @@ add_executable(command_handler_test command_handler_test.cpp)
 configure_target(command_handler_test)
 target_link_libraries(command_handler_test PUBLIC CommandHandler GTest::gtest GTest::gtest_main GTest::gmock GTest::gmock_main)
 add_test(NAME CommandHandlerTest COMMAND command_handler_test)
+
+add_executable(CommandStore_test command_store_test.cpp)
+configure_target(CommandStore_test)
+target_link_libraries(CommandStore_test PUBLIC CommandHandler GTest::gtest GTest::gtest_main GTest::gmock GTest::gmock_main)
+add_test(NAME CommandStoreTest COMMAND CommandStore_test)
