+++ src/agent/configuration_parser/CMakeLists.txt
@@ -11,7 +11,7 @@ set_common_settings()
 find_package(yaml-cpp CONFIG REQUIRED)
 
 add_library(ConfigurationParser src/configuration_parser.cpp src/configuration_parser_utils.cpp)
-target_include_directories(ConfigurationParser PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(ConfigurationParser PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src)
 target_link_libraries(ConfigurationParser PUBLIC yaml-cpp::yaml-cpp Logger PRIVATE Config)
 
 include(../../cmake/ConfigureTarget.cmake)
