+++ src/agent/configuration_parser/src/configuration_parser_utils.cpp
@@ -1,4 +1,4 @@
-#include <configuration_parser_utils.hpp>
+#include "configuration_parser_utils.hpp"
 #include <queue>
 #include <unordered_set>
