+++ src/agent/configuration_parser/tests/CMakeLists.txt
@@ -8,6 +8,6 @@ add_test(NAME ConfigParserTest COMMAND ConfigurationParser_test)
 
 add_executable(ConfigurationParserUtils_test configuration_parser_utils_test.cpp ${CMAKE_CURRENT_SOURCE_DIR}/../src/configuration_parser_utils.cpp)
 configure_target(ConfigurationParserUtils_test)
-target_include_directories(ConfigurationParserUtils_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../include)
+target_include_directories(ConfigurationParserUtils_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../include ${CMAKE_CURRENT_SOURCE_DIR}/../src)
 target_link_libraries(ConfigurationParserUtils_test PUBLIC GTest::gtest GTest::gtest_main GTest::gmock GTest::gmock_main yaml-cpp::yaml-cpp)
 add_test(NAME ConfigParserUtilsTest COMMAND ConfigurationParserUtils_test)
