+++ src/agent/configuration_parser/tests/configuration_parser_utils_test.cpp
@@ -1,4 +1,4 @@
-#include <configuration_parser_utils.hpp>
+#include "configuration_parser_utils.hpp"
 #include <gtest/gtest.h>
 #include <yaml-cpp/yaml.h>
