+++ src/agent/multitype_queue/CMakeLists.txt
@@ -8,8 +8,17 @@ project(MultiTypeQueue)
 include(../../cmake/CommonSettings.cmake)
 set_common_settings()
 
+# MessageEntry target
+
 find_package(nlohmann_json REQUIRED)
-find_package(fmt REQUIRED)
+
+add_library(MessageEntry INTERFACE)
+
+target_include_directories(MessageEntry INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/include/message_entry)
+target_link_libraries(MessageEntry INTERFACE nlohmann_json::nlohmann_json)
+
+# MultiTypeQueue target
+
 find_package(Boost REQUIRED COMPONENTS asio)
 
 add_library(MultiTypeQueue src/storage.cpp src/multitype_queue.cpp)
@@ -20,12 +29,11 @@ target_include_directories(MultiTypeQueue PUBLIC
 target_link_libraries(MultiTypeQueue
         PUBLIC
         Config
-        nlohmann_json::nlohmann_json
+        MessageEntry
         Boost::asio
         Logger
         PRIVATE
-        Persistence
-        fmt::fmt)
+        Persistence)
 
 include(../../cmake/ConfigureTarget.cmake)
 configure_target(MultiTypeQueue)
