+++ src/agent/multitype_queue/tests/CMakeLists.txt
@@ -2,7 +2,6 @@ find_package(GTest REQUIRED)
 
 add_executable(test_MultiTypeQueue multitype_queue_test.cpp)
 configure_target(test_MultiTypeQueue)
-target_include_directories(test_MultiTypeQueue PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../src)
 target_link_libraries(test_MultiTypeQueue PUBLIC
     MultiTypeQueue
     GTest::gtest
@@ -13,7 +12,6 @@ add_test(NAME MultiTypeQueueTest COMMAND test_MultiTypeQueue)
 
 add_executable(test_storage storage_test.cpp)
 configure_target(test_storage)
-target_include_directories(test_storage PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../src)
 target_link_libraries(test_storage
     MultiTypeQueue
     GTest::gtest
