+++ src/agent/persistence/tests/CMakeLists.txt
@@ -2,6 +2,7 @@ find_package(GTest REQUIRED)
 
 add_executable(test_SQLiteManager sqlite_manager_test.cpp)
 configure_target(test_SQLiteManager)
+target_include_directories(test_SQLiteManager PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../src)
 target_link_libraries(test_SQLiteManager PRIVATE
     Persistence
     GTest::gtest
