+++ src/agent/src/command_handler_utils.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
+#include <command_entry.hpp>
 #include <imultitype_queue.hpp>
 #include <moduleWrapper.hpp>
-#include <module_command/command_entry.hpp>
 
 #include <boost/asio/awaitable.hpp>
 #include <nlohmann/json.hpp>
