+++ src/agent/tests/CMakeLists.txt
@@ -11,7 +11,8 @@ endif()
 
 add_executable(agent_registration_test agent_registration_test.cpp)
 configure_target(agent_registration_test)
-target_include_directories(agent_registration_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../src)
+target_include_directories(agent_registration_test PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../src
+    ${CMAKE_CURRENT_SOURCE_DIR}/../agent_info/src)
 target_link_libraries(agent_registration_test PRIVATE Agent GTest::gmock GTest::gtest)
 add_test(NAME AgentRegistrationTest COMMAND agent_registration_test)
