+++ src/modules/inventory/include/inventory.hpp
@@ -13,9 +13,10 @@
 #include <configuration_parser.hpp>
 #include <dbsync.hpp>
 #include <inventoryNormalizer.hpp>
-#include <multitype_queue.hpp>
 #include <sysInfoInterface.hpp>
 
+#include <command_entry.hpp>
+#include <message.hpp>
 #include <moduleWrapper.hpp>
 
 #include <boost/asio/awaitable.hpp>
