+++ src/modules/logcollector/include/logcollector.hpp
@@ -1,7 +1,10 @@
 #pragma once
 
+#include <config.h>
+
+#include <command_entry.hpp>
+#include <message.hpp>
 #include <moduleWrapper.hpp>
-#include <multitype_queue.hpp>
 
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/steady_timer.hpp>
